Revert "moved AbstractWorkflowBasedMDSDBlackboardLaunchConfigurationDelegate to de.uka.ipd.sdq.workflow.mdsd.core"
This reverts commit 75580fd0e79b371d57b40d39804233e5650a2362.
