+++ bundles/de.uka.ipd.sdq.workflow.mdsd/META-INF/MANIFEST.MF
@@ -38,7 +38,6 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.4.0",
  org.eclipse.xtext;bundle-version="2.0.1",
  org.eclipse.xtext.xtext.generator,
  de.uka.ipd.sdq.workflow.workbench,
- de.uka.ipd.sdq.workflow.mdsd.blackboard;visibility:=reexport,
- de.uka.ipd.sdq.workflow.mdsd.core;bundle-version="5.2.0";visibility:=reexport
+ de.uka.ipd.sdq.workflow.mdsd.blackboard;visibility:=reexport
 Bundle-Activator: de.uka.ipd.sdq.workflow.mdsd.Activator
 Bundle-ActivationPolicy: lazy
